+++ Changelog.md
@@ -6,9 +6,22 @@
 
 #### Balance changes
 
+##### General
+
 * [X] Threshold for Inuhh to die in water at the lower border is now more merciful
 * [X] Exchanging items will drop the old item instead of deleting it
 * [ ] Several others (testing)
+
+##### Enemies
+
+* [X] Shilectrons have damage reduction to 1
+* [X] Shilectrons in Shitomium fight will now expand their radius and fluctuate if attacked
+* [X] Raised Shilectrons will eject sparks and revert to normal state
+* [X] Shitomium will revert Shilectron directions in Hard as well
+
+##### Levels
+
+
 
 #### Aesthetic changes
 
@@ -18,6 +31,7 @@
 #### Bugfixes
 
 * [X] Fixed and improved credits
+* [X] Shitomium will now automatically give its dropped key to Inuhh
 * [X] Exiting level on new save now leads back to the world map
 
 #### Code quality
@@ -30,6 +44,8 @@
 ### Version 1.0.1
 
 #### Balance changes
+
+##### Levels
 
 * Fixed unfair Shi coin in 5-1-2